server/requirements.txt
fastapi==0.63.0
uvicorn==0.13.2
google-api-python-client==1.12.8
requests~=2.25.1
google-cloud-storage==1.35.0
python-multipart==0.0.5
aiofiles==0.4.0